# Efficient Fiat-Shamir challenges using hashing