By avoiding ssz merklelization, we significantly reduce the latency of generating FS
challenges.

```
name                 old time/op    new time/op    delta
VerifyMultiple/8-6     73.9ms ± 0%    40.7ms ± 0%   ~     (p=1.000 n=1+1)
VerifyMultiple/16-6     138ms ± 0%      81ms ± 0%   ~     (p=1.000 n=1+1)

name                 old alloc/op   new alloc/op   delta
VerifyMultiple/8-6     25.2MB ± 0%    22.1MB ± 0%   ~     (p=1.000 n=1+1)
VerifyMultiple/16-6    50.4MB ± 0%    44.1MB ± 0%   ~     (p=1.000 n=1+1)

name                 old allocs/op  new allocs/op  delta
VerifyMultiple/8-6       400k ± 0%      301k ± 0%   ~     (p=1.000 n=1+1)
VerifyMultiple/16-6      800k ± 0%      603k ± 0%   ~     (p=1.000 n=1+1)
```