@roberto-bayardo 2022-08-22
This is great! I assume this is not a backwards compatible change since the hash values presumably change?