@roberto-bayardo 2022-08-22
OK, will try to implement in prysm soon then.